requirements.txt
torch
torchvision
transformers
pandas
numpy
scikit-learn
matplotlib
Pillow
requests
